COT-2: fixed null handling in configuration getters for primitive types

+++ java-me-smartrest/java-me-smartrest-client/src/main/java/com/cumulocity/me/agent/integration/impl/IntegrationHandler.java
@@ -18,7 +18,7 @@ public class IntegrationHandler {
 
     public IntegrationHandler(ExternalIdProvider idProvider, SmartHttpConnection connection, ConfigurationService configService) {
         this.idProvider = idProvider;
-        this.requiredInterval = configService.getInt(ConfigurationKey.AGENT_MONITORING_REQUIRED_INTERVAL);
+        this.requiredInterval = configService.getInt(ConfigurationKey.AGENT_MONITORING_REQUIRED_INTERVAL).intValue();
         this.connection = connection;
     }
